+++ CHANGELOG.md
@@ -1,13 +1,15 @@
-## [1.1.0] - 2025-09-22
+## [1.1.0] - 2025-09-24
 
 ### New Features
 
 - DNS zone and records configuration
 - Virtual Machines A and CNAME records creation (root)
+- Resource group default domain creation (with base DNS domain)
 
 ### Improvements
 
 - [BREAKING] Refactored resource group resource configuration (root)
+- [BREAKING] Changed context order
 
 ## [1.0.0] - 2025-09-15